generated yet other optim files for full IK project
another approach is taken with computeIiwaSelfMotionAdmissibleRanges.m where the admissible swivel ranges, for a given position and orientation of the robot base frame, is computed based on jlimits, force capa, deflection of the structure. The code works but is rather slow.. needs vectorization

created selfMotionInspector.m which is a helper class that set iiwa config on vrep based on the corresponding clicked swivel.